+++ pdfplumber/table.py
@@ -8,16 +8,20 @@ DEFAULT_MIN_WORDS_VERTICAL = 3
 DEFAULT_MIN_WORDS_HORIZONTAL = 1
 
 
-def snap_edges(edges, tolerance=DEFAULT_SNAP_TOLERANCE):
+def snap_edges(
+    edges, x_tolerance=DEFAULT_SNAP_TOLERANCE, y_tolerance=DEFAULT_SNAP_TOLERANCE
+):
     """
     Given a list of edges, snap any within `tolerance` pixels of one another
     to their positional average.
     """
-    v, h = [list(filter(lambda x: x["orientation"] == o, edges)) for o in ("v", "h")]
+    by_orientation = {"v": [], "h": []}
+    for e in edges:
+        by_orientation[e["orientation"]].append(e)
 
-    snap = utils.snap_objects
-    snapped = snap(v, "x0", tolerance) + snap(h, "top", tolerance)
-    return snapped
+    snapped_v = utils.snap_objects(by_orientation["v"], "x0", x_tolerance)
+    snapped_h = utils.snap_objects(by_orientation["h"], "top", y_tolerance)
+    return snapped_v + snapped_h
 
 
 def join_edge_group(edges, orientation, tolerance=DEFAULT_JOIN_TOLERANCE):
@@ -47,7 +51,7 @@ def join_edge_group(edges, orientation, tolerance=DEFAULT_JOIN_TOLERANCE):
     return joined
 
 
-def merge_edges(edges, snap_tolerance, join_tolerance):
+def merge_edges(edges, snap_x_tolerance, snap_y_tolerance, join_tolerance):
     """
     Using the `snap_edges` and `join_edge_group` methods above,
     merge a list of edges into a more "seamless" list.
@@ -59,8 +63,8 @@ def merge_edges(edges, snap_tolerance, join_tolerance):
         else:
             return ("v", edge["x0"])
 
-    if snap_tolerance > 0:
-        edges = snap_edges(edges, snap_tolerance)
+    if snap_x_tolerance > 0 or snap_y_tolerance > 0:
+        edges = snap_edges(edges, snap_x_tolerance, snap_y_tolerance)
 
     if join_tolerance > 0:
         _sorted = sorted(edges, key=get_group)
@@ -412,6 +416,8 @@ DEFAULT_TABLE_SETTINGS = {
     "explicit_vertical_lines": [],
     "explicit_horizontal_lines": [],
     "snap_tolerance": DEFAULT_SNAP_TOLERANCE,
+    "snap_x_tolerance": None,
+    "snap_y_tolerance": None,
     "join_tolerance": DEFAULT_JOIN_TOLERANCE,
     "edge_min_length": 3,
     "min_words_vertical": DEFAULT_MIN_WORDS_VERTICAL,
@@ -475,6 +481,8 @@ class TableFinder(object):
         for var, fallback in [
             ("text_x_tolerance", "text_tolerance"),
             ("text_y_tolerance", "text_tolerance"),
+            ("snap_x_tolerance", "snap_tolerance"),
+            ("snap_y_tolerance", "snap_tolerance"),
             ("intersection_x_tolerance", "intersection_tolerance"),
             ("intersection_y_tolerance", "intersection_tolerance"),
         ]:
@@ -573,10 +581,15 @@ class TableFinder(object):
 
         edges = list(v) + list(h)
 
-        if settings["snap_tolerance"] > 0 or settings["join_tolerance"] > 0:
+        if (
+            settings["snap_x_tolerance"] > 0
+            or settings["snap_y_tolerance"] > 0
+            or settings["join_tolerance"] > 0
+        ):
             edges = merge_edges(
                 edges,
-                snap_tolerance=settings["snap_tolerance"],
+                snap_x_tolerance=settings["snap_x_tolerance"],
+                snap_y_tolerance=settings["snap_y_tolerance"],
                 join_tolerance=settings["join_tolerance"],
             )
         return utils.filter_edges(edges, min_length=settings["edge_min_length"])